requirements.txt
azure_storage==0.37.0
cssutils==2.9.0
openai==0.27.4
pandas==1.5.3
python-dotenv==1.0.0
requests==2.31.0
streamlit==1.28.2
